[add]コンポーネントのファイル分け
