+++ src/component/Header.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+class Header extends React.Component {
+  render() {
+    return <h1 className="header">ゆめみ フロントエンドコーディング試験</h1>;
+  }
+}
+
+export default Header;
